+++ nodes/Auvik/resources/configuration/description.ts
@@ -39,7 +39,7 @@ export const configurationFields: INodeProperties[] = [
     displayOptions: { show: { resource: ['configuration'], operation: ['getMany'] } },
   },
   { displayName: 'Device ID', name: 'filterDeviceId', type: 'string', default: '', displayOptions: { show: { resource: ['configuration'], operation: ['getMany'] } } },
-  { displayName: 'Backup Time Preset', name: 'backupTimePreset', type: 'options', default: 'LAST_30_DAYS', options: [...dateTimePresetOptions, { name: 'Custom', value: 'CUSTOM' }], displayOptions: { show: { resource: ['configuration'], operation: ['getMany'] } } },
+  { displayName: 'Backup Time Preset', name: 'backupTimePreset', type: 'options', default: 'NO_FILTER', options: [...dateTimePresetOptions, { name: 'Custom', value: 'CUSTOM' }], displayOptions: { show: { resource: ['configuration'], operation: ['getMany'] } } },
   { displayName: 'Backup Time After', name: 'filterBackupTimeAfter', type: 'string', default: '', description: 'YYYY-MM-DDTHH:mm:ss.SSS[Z]', displayOptions: { show: { resource: ['configuration'], operation: ['getMany'], backupTimePreset: ['CUSTOM'] } } },
   { displayName: 'Backup Time Before', name: 'filterBackupTimeBefore', type: 'string', default: '', description: 'YYYY-MM-DDTHH:mm:ss.SSS[Z]', displayOptions: { show: { resource: ['configuration'], operation: ['getMany'], backupTimePreset: ['CUSTOM'] } } },
   {
@@ -49,8 +49,8 @@ export const configurationFields: INodeProperties[] = [
     default: 'ALL',
     options: [
       { name: 'All', value: 'ALL' },
-      { name: 'Running Only', value: 'RUNNING' },
       { name: 'Not Running Only', value: 'NOT_RUNNING' },
+      { name: 'Running Only', value: 'RUNNING' },
     ],
     displayOptions: { show: { resource: ['configuration'], operation: ['getMany'] } },
   },